Stefan comments
